requirements.txt
streamlit
google-generativeai
groq
huggingface_hub
python-dotenv
google-search-results